add a screenreader only describedby div
This allows non visual users to have a description of the fields they
are using. This is extremely helpful for understanding the context in
which you're working. Because everything is being passed in, the
consuming app can define exactly how they would like to convay their
field. It also allows them to easily drop in any the appropriate text
for the appropriate translation.